requirements.txt
setuptools
greenlet
bottle
gevent
gevent-websocket
websocket-client
requests

# for folks running raspberry pis
# we have no proof of anyone using another board yet, but when that 
# happens, you might want to comment this out.
RPi.GPIO

# List of all supported adafruit modules for thermocouples
adafruit-circuitpython-max31855
adafruit-circuitpython-max31856

# for folks using sw spi (bit banging)
adafruit-circuitpython-bitbangio

# untested - for PT100 platinum thermocouples
#adafruit-circuitpython-max31865

# untested - for mcp9600 and mcp9601
#adafruit-circuitpython-mcp9600

# Google Sheets API dependencies
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
